Improved bench code related to transaction overhead benchmarking.
Now the benchmark is done in same conditions for clojure.jdbc
and java.jdbc related to transactions.

In previous benchmark code, the transaction strategy used for
clojure.jdbc it was more powerfull, and additional features
implies some additional overhead.
Now basic transaction strategy is used for the benchmark, that
imitates the java.jdbc behavior.